fix: 순환참조 제거
